Narrow mdlFile 0033e540 pointer parameter

+++ src/Event/Fcl/fclCombineEffect.c
@@ -76,8 +76,8 @@ u8 FUN_00419f20(int param_1,int *param_2);
 u32 FUN_0041a2c0(int param_1,int *param_2);
 u32 FUN_0041a310(u64 param_1,int param_2);
 u8 FUN_0041a3f0(u64 param_1,int param_2);
-u8 FUN_0041a730(int param_1,int param_2);
-u8 FUN_0041a900(u64 param_1,int param_2);
+u8 FUN_0041a730(u64 param_1,int param_2);
+u8 FUN_0041a900(int param_1,int param_2);
 u8 FUN_0041abc0(int param_1,int *param_2);
 u32 FUN_0041ae20(int param_1);
 u8 FUN_0041aff0(int param_1,int param_2);
@@ -2806,7 +2806,7 @@ u8 FUN_0041a3f0(u64 param_1,int param_2)
 // FUN_0041A730 NONMATCHING
 
 
-u8 FUN_0041a730(int param_1,int param_2)
+u8 FUN_0041a730(u64 param_1,int param_2)
 
 
 
@@ -2874,7 +2874,7 @@ u8 FUN_0041a730(int param_1,int param_2)
 // FUN_0041A900 NONMATCHING
 
 
-u8 FUN_0041a900(u64 param_1,int param_2)
+u8 FUN_0041a900(int param_1,int param_2)
